requirements.txt
PyMySQL==0.9.3
bcrypt==3.1.7
PyJWT==1.7.1
jsonschema==3.2.0
boto3==1.10.4
